A file with many authors
Assuming the following had a hand in creating many of the abbreviations here:
Delta Tao, Noivad, Gorvin, Raul, DragonHawk, Lorikeet, Skirwan? and many more.
Later editions made to allow for customization of certain commands based on variables in the character's file for RPG purposes and to avoid having to keep code in your character file.