The spammy macro
This is the one that tends to spam the log with who it tried to thank. It's used in conjunction with idle.macro. Although I tried to turn off the spammy nature of it, it still spams. At least you can randomly thank people. It does mess with cadding at times. Last updated 20210227